Adjusted file paths and changed stlying for the student profile card
