+++ attendance-tracking-tool-app/src/App.js
@@ -1,0 +1,16 @@
+import React from "react";
+import './App.css';
+import StudentProfile from "./components/StudentProfile.js";
+
+function App() {
+  return (
+    <div className="App">
+      
+
+       {/* Render the StudentProfile page */}
+       <StudentProfile />
+    </div>
+  );
+}
+
+export default App;
