+++ attendance-tracking-tool-app/src/assets/styles/UserProfileCard.css
@@ -2,9 +2,10 @@
 .user-profile-card {
     border: 1px solid #ddd;
     border-radius: 8px;
+    border-color: #333;
     padding: 16px;
     max-width: 300px;
-    background-color: #f9f9f9;
+    background-color: #ffffff;
     box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1);
   }
